Type orm

+++ docker/fclean_docker.sh
@@ -1,6 +1,21 @@
-docker stop $(docker ps -a -q)
-docker rm $(docker ps -a -q)
-docker rmi $(docker images -a -q)
-docker volume rm $(docker volume ls -q)
-docker system prune -a --volumes
+#!/bin/sh
+
+if [ -n "$(docker ps -a -q)" ]; then
+    # shellcheck disable=SC2046
+    docker stop $(docker ps -a -q)
+    # shellcheck disable=SC2046
+    docker rm $(docker ps -a -q)
+fi
+if [ -n "$(docker images -a -q)" ]; then
+    # shellcheck disable=SC2046
+    docker rmi $(docker images -a -q)
+fi
+if [ -n "$(docker volume ls -q)" ]; then
+    # shellcheck disable=SC2046
+    docker volume rm $(docker volume ls -q)
+fi
+
+docker system prune -f -a --volumes
+
+
